requirements.txt
importlib-metadata==4.13.0
celery==5.2.7
Django==3.2.16
djangorestframework==3.14.0
google-api-python-client==2.69.0
redis==4.4.0
requests==2.28.1
django-celery-beat==2.4.0
